Update source.asm
Finishes the code for the function that allows the program to read an integer from the user via the console. The read_int subroutine will read the user's input and store the result in the $v0 register then returns control to the calling function. Also begins the code for the main program and changes the wording for the first prompt on line 2.